pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.6.3",
    "googletrans>=4.0.2",
    "python-dotenv>=1.1.1",
]
